+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - PIM-10870: Fix display of permissions when empty
 - PIM-10860: [SLA] Announcements aren't shown
 - PIM-10745: Fix history display for product's quantified association
+- PIM-10874: fix labels api type consistency
 
 ## Improvements